Move priming logic to AsyncEstimatedTimetableSource
